Revamped the method that adds declarations to the symtab
This version avoids creating intermediate data structures and instead
adds each declaration to the symbol table using the put method.  When
the symbol table entries are needed for the parameter list, it
retrieves the declarations from the symbol table using the get()
method.  This also removes the name mangling code, which is done by
the symbol table.