Maintain the complement of the multiverse
The complement is the set of configuration not covered by any element
in the multiverse.  Maintaining allows users to quickly retrieve the
complement and also allows the object to make sure only
mutually-exclusive elements are added to the multiverse, desirable to
prevent accidentally having multiple states for the same
configuration.