Write typedefes to top of scope
Since structs can use typedefs, we also need to move these to the top
of the scope.  Type declarations can't use variables (dependent
types), since this is not supported in C, so it should be safe to move
all type declarations to the top of the scope, as long as the ordering
is preserved.