Use the z3 expression for printing static conditions; save renamings
This prints out the __static_condition_renaming for each static
conditional.  In the future, the static condition emission can be
improved by walking the z3 expression and creating a single renaming
for common predicates.

Also fixes a bug where statement and declaration uses of emitError did
not emit a semi-colon after it, since emitError is just an expression.