Finish function definition desugaring
This uses a reworking of compound statement where the function
definition does not worry about braces or scope itself.

This change also supports recursive functions by correctly handling
scope.  It adds the function to the symtab just after the prototype so
it can be used (in all configurations) inside of the body of the
function, which now just concatenates strings across all
configurations.