Implement the forward struct reference desugaring struct
This feature accounts for forward references to struct that have not
been defined yet.  Since the struct may end up being multiply-defined,
the reference needs to be duplicated. But since it's a forward
reference, we don't know (in a single-pass) whether that's the case.
Instead, we define the forward reference to be the union of all
configurations of the struct and then update the field selection to
use follow this indirection.