Implement anonymous structs the same as named structs
The only difference is that the anonymous struct gets a fresh name and
has no way to be referenced or redefined.

This also adds todo reminders to check for anonymous fields in
structs/unions, e.g., an anonymous union.  These don't need storing in
the lookaside symtab, since they can't be referenced, but they do need
to be desugared.