Fix behavior of conditional inside tracking
Take a production A ::= B C.  If B or C have a conditional inside,
then transitively, A also has a conditional inside.  If C has a
conditional after, then A also has a conditional after, but this is
not the case for B, since the follow set of B is not (necessarily) in
the follow set of C.

The current code does not update conditional inside for A when C has a
conditional inside.  This commit is meant to fix that.