API to get production inside conditional (still testing)
Support to get the production that is surrounded by a conditional node after subparser merging has been added in ForkMergerParser.java - still in testing phase. FMP also now tags list nodes with a new property so that when retrieving production inside a list node with conditional node inside it is handled differently.