Allow FunctionPrototype to be a complete syntactic unit
This is meant to address a bug where the AST has several versions of
the same function prototype and the desugarer does not properly rename
one of them.  The proposed fix is indirect: it allows the function
prototype to be a complete synatctic unit, then combines each
function prototype variation with each body variation, rather than
trying to get the parser to lift the conditional around the entire
function definition.

This is an example that caused the errant desugaring:

    void b(
    #ifdef A
           int a
    #endif
    ) {
    #if defined(A) && defined(B)
      int c = 2;
    #endif
      int d;
    }