SuperC: sourcelinePC: switch to tree representation
Instead of line-range : pc mapping, represent conditional blocks as a
tree. This allows user to do certain types of walks on the tree to get
explicit information on how condition blocks are grouped and organized.

Some term definitions are as follows:
* ConditionalBlock: Conditional block started with a conditional START or NEXT
token and ended with a conditional NEXT/END token.
* ConditionalBlockGroup: Group of sibling conditional blocks which start
with a conditional START token and end with a conditionl END token.