Adding support to handle RANGE tokens scenarios in lexing stage
PPNUM and INT-RANGE-INT patterns overlap, causing the lexer to create the token PPNUM when it encounters a P4 range value like 1..5 instead of producing INT, RANGE, INT tokens.
So making an exception where if lexer encounters a number followed by two periods, it converts the number before the periods as integers, the two periods as RANGE token, and the following integer as an integer token.
Similar for FLOAT_CONSTANTS the lexer ensures that the integer is followed by only one period (if followed by a period). Iff it is followed by one period it is treated as float constants, else it is either a PPNUM and INT-RANGE-INT.