Adding variables to track all productions and to emit after-set under conditional granularity
The "trackAllProductions" will be used to indicate whether to track all productions under conditional granularity or not (else the specific C productions should be used).
The "emitConditionalAfterSet" will be used to emit the conditional after-set (of productions) when the conditional inside-set is not available. This was first used when gathering statistics about productions inside conditionals in P4 programs where some productions were not emitted because they were in the after set and not in the inside set. There are some cases where the production(s) is/are inside both sets, so conditional after-set should only be emitted when conditional inside-set is empty.
The "conditionGranularity" is a new overload function will be used to set the trackAllProductions and emitConditionalAfterSet flags along with conditionGranularity flag. Will be used initially with SuperP4 (in the SuperP4.java file).