separating assignment and method call && lvalues + lvalue support in call graph
separating out assignment and method call constructs in the parser for easier identification in the call graph generation. Plus now lvalue is supported in call graph generation with a single assertion that method calls' lvalue only contain name and no dot values. Still yet to assert that lvalues called under assignment are only of the three constructs mentioned in comments