Remove backwards dependency on cdesugar.CContext from core
ForkMergeParser depended on cdesugarer.CContext, because
cdesugarer.CContext needed access to the parse error condition.  But
this imposes a backwards dependency.  Instead, this exposes the parser
error condition via getErrorCond() from the parser and now
cdesugarer.CContext calls it to retrieve it.  This design still
depends on some static field in CContext that should probably be
addressed later.

This change also adds the setNewErrorHandling and setSaveErrorCond
methods to the parser and uses them in SuperC and SugarC.