(Not fully tested) storing parsed type parameters on node & specialized type uses function prototype's type parameters
checksum-l4-bmv2.p4 test case points out a case where the lookahead function is invoked multiple times but with different type parameters (hence the return value is different), the old implementation just appended all the different type parameters from different invocations, so it was just retrieving the first value at all invocations and not the respective passed in type parameter since it was just an arraylist. So idea is to store it on the node and not in the object. Does break cases like generic-struct.p4 ; need to fix it.

SpecializedType class now uses function prototype as it was previously unaccounted for.