Support for Generator Constructs for Constructs with Generic Types
1. New subclasses are introduced to differentiate between constructs that contain generic types and constructs that have all values’ types defined (no generic types)
a. Type parameters are specific to generator constructs while regular parameters are still generalized to all constructs
b. The generator class takes in that construct’s regular version (ExternDeclarationGenerator takes in ExternDeclaration object) -> an adapter design so all logic is inherited from the regular construct and is not duplicated in two places
c. The generateInstance function takes in the passed in parameters and type parameters and mappings for those values, replaces all instance of the generic types with the passed in value, creates a similar object of regular type with no generic values and returns that
I. If some generic types were not passed in, then the newly created instance is also a generator instance but containing only the generics that were not substituted (e.x. Construct containing child functions was instantiated but child function has generic that is not that of the parent, so the function is still a generator function since the child function’s generic values were not defined when the parent construct was instantiated)
II. Symtab look up function is changed so that if a generator construct is fetched, the generatorInstance method is called on that construct to retrieved a construct whose generic values are substituted with the passed in values
III. Note: note all generic related functionalities have been changed in this commit. Only certain constructs are modified to support the new generic-suported model (functionPrototype, methodCallStatements, and instantiation functions are supported at this point). This is to test the implementation with an example program and then expand to all constructs (used test case: https://github.com/p4lang/p4c/blob/main/testdata/p4_16_samples/extern3.p4)
IV. The class TemporaryParameterValues is introduced to keep track of passed in parameter values. The parsed/passed in type arguments and parameters are stored in a stack of TemporaryParameterValues to account for nested cases
V. Since method calls are evaluated in the type checking stage itself, the final call method is stored in the node (using setProperty function ) so that the call graph visitor method doesn’t have to recalculate the callee construct.