Fixing specializedType logic when it is called by instantiation
specializedType and headerStackType:
instantiation can use either of these two consturcts as the type its initializing. But instantiation provides the parameter arguments while specializedType provides the type arguments. BUT specializedType is not exclusive to instantiation, so it can be invoked from other places. So it can have its own temp value scope but so can instantiation parameter values, so manually override not adding a new temporary value scope for both those constructs.