Fixing bug where functionPrototype wasn't tagged as generator
Function prototype was not converted to a generator function when the return value wasn't of generic type and no new optTypeParameters were passed in. If the type parameter was passed in through a parent construct - like extern declaration - then function prototype remained normal class. Fixed that by checking if any parameter is of generic type. Algorithm can be optimized