Cleaner implemented to resolve generics - refactored a lot
Handling of generics had a big bug - so refactored the code to handle generics better. Now when resolving generic constructs we go through the AST again for the body of the construct we are resolving generics for. Required more values to be passed and hence the change. Plus updating test cases