Adding support to parse multiple final instantiations for macro collections
There are cases when a P4 program invokes the package/switch multiple times with different configurations (there might be two parsers defined for example, and one switch/package uses parser A while another uses parser B). So adding support so that macro usage output will be displayed for any package or switch declaration that matches the passed in template (the -preprocessorUsageMatrix -templateFileForMatrix flags).